Updated Before Publishing
Updated And Finalized Before Publishing With proper usage of allowing one to change the input file name and output file name that it is using!
To truly create a universal plugin class for developers to utilize the new built in nodeJS way to bundle and compile executables! via a post inject method!